requirements.txt
decord==0.3.9
jsonnet==0.15.0
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
parso==0.5.2
PasteDeploy==2.1.0
pbkdf2==1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
plaster==1.0
plaster-pastedeploy==0.7
prompt-toolkit==3.0.2
protobuf==3.11.3
psutil==5.6.7
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.19
Pygments==2.5.2
pyhocon==0.3.54
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyramid==1.10.4
pyramid-mailer==0.15.1
PySocks==1.7.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.2
repoze.sendmail==4.4.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel-yaml==0.15.87
six==1.12.0
soupsieve==1.9.5
SQLAlchemy==1.3.16
tensorboard==2.1.1
torch==1.6.0
torchvision==0.7.0
tqdm==4.36.1
traitlets==4.3.3
transaction==3.0.0
translationstring==1.3
typed-args==0.3.5
urllib3==1.24.2
velruse==1.1.1
venusian==3.0.0
wcwidth==0.1.7
WebOb==1.8.6
Werkzeug==1.0.0
WTForms==2.2.1
wtforms-recaptcha==0.3.2
zope.deprecation==4.4.0
zope.interface==5.1.0
zope.sqlalchemy==1.3
